Changes to use an ssp_blog macro and to properly set the IP
Prior to this change, there was a #define that redefined blog to prepend
[obs-ssp] to the log message.  However, it used C style string
concatination which make it impossible to add a program-built string to
log.  To facilitate generic logging, I'm not using ssp_blog to log ssp
messages, freeing the blog to be used raw by other code.

The IP address was not properly set before setStream was called.  It
caused the camera to not start when obs would start, because the first
setStream method would fail because there was no IP address set in the
url.  We now set the IP address in ssp_source_update prior to calling
setStream.

Finally, to facilitate additional debug information reporting, the
setStream methods now take a StatusReasonUpdateCallback argument which
is a callback that also includes a QString stating why something
succeeded or failed.