ssp: isolate libssp to standalone process
Since we can't find out why it always crash when libssp loaded on
obs process, we isolate ssp protocol to a standalone process called
ssp-connector and use stdout to communicate with obs.

And it happends that we can use the benefit of Rosetta2 to let
obs on Apple Silicon support obs-ssp natively by only running
ssp-connector process on Rosetta2. It's only a network thread
so it's easy for Roestta2 to achieve its performance need.

Signed-off-by: Yibai Zhang <xm1994@gmail.com>